requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.11.0=pyhd3eb1b0_1
astor=0.8.1=py36hecd8cb5_0
blas=1.0=openblas
bzip2=1.0.8=h1de35cc_0
c-ares=1.17.1=h9ed2024_0
ca-certificates=2021.1.19=hecd8cb5_0
cairo=1.14.12=hc4e6be7_4
certifi=2020.12.5=py36hecd8cb5_0
cycler=0.10.0=py36hecd8cb5_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.1=ha9ee91d_0
freetype=2.10.4=ha233b18_0
gast=0.4.0=py_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.66.1=h9bbe63b_0
graphite2=1.3.14=h38d11af_0
grpcio=1.31.0=py36h7580e61_0
h5py=2.8.0=py36h878fce3_3
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h0a44026_3
importlib-metadata=2.0.0=py_1
jasper=2.0.14=h636a363_1
jpeg=9b=he5867d9_2
keras=2.2.4=0
keras-applications=1.0.8=py_1
keras-base=2.2.4=py36_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py36h23ab428_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopenblas=0.3.10=h0794777_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.14.0=h2842e9f_0
libtiff=4.1.0=hcb84e12_1
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
lz4-c=1.9.3=h23ab428_0
markdown=3.3.3=py36hecd8cb5_0
matplotlib=3.1.1=py36h54f8f79_0
mock=4.0.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nomkl=3.0=0
numpy=1.19.2=py36h63973fd_0
numpy-base=1.19.2=py36h68fea81_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1i=h9ed2024_0
pandas=1.1.3=py36hb1e8313_0
pcre=8.44=hb1e8313_0
pillow=6.1.0=py36hb68e598_0
pip=20.3.3=py36hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
protobuf=3.14.0=py36h23ab428_1
py-opencv=3.4.2=py36h7c891bd_1
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.6.10=hf48f09d_2
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py36h9ed2024_1
readline=8.1=h9ed2024_0
scipy=1.5.2=py36hadf427e_0
setuptools=52.0.0=py36hecd8cb5_0
six=1.15.0=py36hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
tensorboard=1.13.1=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-base=1.13.1=mkl_py36hc36dc97_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=1.1.0=py36_1
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py36h9ed2024_0
tqdm=4.48.2=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py36h1de35cc_1
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
